backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
xlsxwriter==3.2.5
